feature/DEV-FIX

+++ src/main/java/be/vdab/scrumjava202504/products/SimpleProductDTO.java
@@ -19,7 +19,7 @@ public class SimpleProductDTO {
         return ean;
     }
 
-    public int getProductId() {
+    public long getProductId() {
         return productId;
     }
 }
